Extend the auto-generation capability with custom logic for OU move
